Change template vector of size<2> to template<N>
In a template class a vector is declared of what appears to be a fixed
size of 2 dimensions. Test code declares the template with 3 dimensions,
which causes errors when pushing values onto this vector.

It is not that clear to me that this is correct. This should work if
using data of 2 dimensions, but that may be a coincidence. That the
vector is populated using an N-bounded loop for the neighbor_index
suggests that this is a correct solution.